pyproject.toml
[project]
name = "edaijiamcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
]
